// SPA/src/app/models/alg-gen.ts
export default interface AlgGen {
  servicoVeiculo: string;
  novasGeracoes: number;
  populacoesEstaveis: number;
  dimensaoPopulacao: number;
  probabilidadeCruzamento: number;
  probabilidadeMutacao: number;
  percentagemIndividuosMelhores: number;
  tempoAbsolutoParagem: number;
  avaliacaoTermino: number;
}